chore:라우팅 설정 변경
